Rename the Ruby namespace from Spreen::Wiki to SpreenWiki in the gem sources
