+++ RubyGem/exe/spreen
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
 # frozen_string_literal: true
 
-require 'spreen/wiki'
+require 'spreen_wiki'
 
-exit Spreen::Wiki::CLI.start
+exit SpreenWiki::CLI.start
