+++ RubyGem/lib/spreen_wiki/templates/category/english.md
