+++ RubyGem/lib/spreen_wiki/templates/category/japanese.md
