+++ RubyGem/lib/spreen_wiki/templates/owner/japanese.md
