+++ RubyGem/lib/spreen_wiki/version.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 # rbs_inline: enabled
 
-module Spreen
-  module Wiki
-    VERSION = '0.1.0'
-  end
+module SpreenWiki
+  VERSION = '0.1.0'
 end
